+++ lib/generics/tests/hash_table/ht_sanity.c
@@ -90,26 +90,25 @@ void str_destroy(void *str) {
 }
 
 // unit tests
-struct hash_table *before(char **keys, size_t keys_size, struct string *strings,
-                          size_t strings_size) {
+struct hash_table *before(char **keys, size_t keys_size, struct string *strings, size_t strings_size) {
   assert(keys_size == strings_size);
   struct hash_table *table = table_init(cmpr_keys, NULL, NULL);
   assert(table);
   assert(table->num_of_elements == 0);
 
   for (size_t i = 0; i < keys_size; i++) {
-    struct string *old = table_put(table, keys[i], strlen(keys[i]) + 1,
-                                   &strings[i], sizeof strings[i]);
+    struct string *old = table_put(table, keys[i], strlen(keys[i]) + 1, &strings[i], sizeof strings[i]);
     assert(!old);
   }
 
   return table;
 }
 
-void after(struct hash_table *table) { table_destroy(table); }
+void after(struct hash_table *table) {
+  table_destroy(table);
+}
 
-void table_put_empty_table_test(char **keys, size_t keys_size,
-                                struct string *strings, size_t strings_size) {
+void table_put_empty_table_test(char **keys, size_t keys_size, struct string *strings, size_t strings_size) {
   // given
   struct hash_table *table = before(keys, keys_size, strings, strings_size);
 
@@ -120,19 +119,15 @@ void table_put_empty_table_test(char **keys, size_t keys_size,
   after(table);
 }
 
-void table_put_with_replace_value_test(char **keys, size_t keys_size,
-                                       struct string *strings,
-                                       size_t strings_size) {
+void table_put_with_replace_value_test(char **keys, size_t keys_size, struct string *strings, size_t strings_size) {
   // given
   struct hash_table *table = before(keys, keys_size, strings, strings_size);
 
   char *another_str = "not the original string";
-  struct string another_string_same_key = {.str_size = strlen(another_str),
-                                           .str = another_str};
+  struct string another_string_same_key = {.str_size = strlen(another_str), .str = another_str};
   // when
   struct string *old =
-      table_put(table, keys[0], strlen(keys[0]), &another_string_same_key,
-                sizeof another_string_same_key);
+    table_put(table, keys[0], strlen(keys[0]), &another_string_same_key, sizeof another_string_same_key);
   // then
   assert(old);
   assert(cmpr_values(old, strings) == 0);
@@ -155,8 +150,7 @@ void table_put_with_resize_test(size_t size) {
 
   // when
   for (size_t i = 0; i < size; i++) {
-    table_put(table, keys[i], strlen(keys[i]) + 1, &strings[i],
-              sizeof strings[i]);
+    table_put(table, keys[i], strlen(keys[i]) + 1, &strings[i], sizeof strings[i]);
   }
 
   // then
@@ -172,19 +166,16 @@ void table_put_with_resize_test(size_t size) {
   destroy_keys(keys, size);
 }
 
-void table_remove_test(char **keys, size_t keys_size, struct string *strings,
-                       size_t strings_size) {
+void table_remove_test(char **keys, size_t keys_size, struct string *strings, size_t strings_size) {
   // given
   struct hash_table *table = before(keys, keys_size, strings, strings_size);
 
-  unsigned long long old_size = table_size(table);
+  size_t old_size = table_size(table);
 
   // when
   struct string *first = table_remove(table, keys[0], strlen(keys[0]) + 1);
-  struct string *last =
-      table_remove(table, keys[keys_size - 1], strlen(keys[keys_size - 1]) + 1);
-  struct string *mid =
-      table_remove(table, keys[keys_size / 2], strlen(keys[keys_size / 2]) + 1);
+  struct string *last = table_remove(table, keys[keys_size - 1], strlen(keys[keys_size - 1]) + 1);
+  struct string *mid = table_remove(table, keys[keys_size / 2], strlen(keys[keys_size / 2]) + 1);
 
   // then
   assert(first && last && mid);
@@ -201,8 +192,7 @@ void table_remove_test(char **keys, size_t keys_size, struct string *strings,
   after(table);
 }
 
-void table_get_test(char **keys, size_t keys_size, struct string *strings,
-                    size_t strings_size) {
+void table_get_test(char **keys, size_t keys_size, struct string *strings, size_t strings_size) {
   // given
   struct hash_table *table = before(keys, keys_size, strings, strings_size);
